renamed extractor module
