+++ leptos_i18n_parser/src/lib.rs
@@ -3,7 +3,18 @@
 #![allow(clippy::too_many_arguments, clippy::result_unit_err)]
 
 pub mod error;
-pub mod extraction;
+mod extractor;
 pub mod formatters;
-pub mod parser;
+mod parser;
 pub mod utils;
+
+pub use extractor::extract_locales;
+pub use parser::parse_locales_raw;
+
+pub mod parsing {
+    pub use super::parser::{RawParsedLocales, parse_locales_raw};
+}
+
+pub mod extraction {
+    pub use super::extractor::{ParsedLocales, defaults::DefaultedLocales, extract_locales};
+}
